+++ pkg/apis/deployment/definitions.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Adam Janikowski
 //
 
 package deployment
@@ -39,5 +37,6 @@ const (
 )
 
 var (
-	ArangoDeploymentShortNames = []string{"arangodb", "arango"}
+	ArangoDeploymentShortNames             = []string{"arangodb", "arango"}
+	ArangoClusterSynchronizationShortNames = []string{"arangoclustersync"}
 )