+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add ArangoJob and Apps Operator
 - Use Go 1.17
 - Add metrics for the plan actions
+- Add ArangoClusterSynchronization Operator
+- Update licenses
 
 ## [1.2.6](https://github.com/arangodb/kube-arangodb/tree/1.2.6) (2021-12-15)
 - Add ArangoBackup backoff functionality